Multi-element copy optimisation for Permutation._get_perm()
Minor edit to allow implementation of __slots__ in slowseq

Permutation._get_perm() now senses zero runs in the tree path, and recognises them as blocks of unpermuted elements. These blocks are copied to out using extend() rather than append(). This should reduce the time to build terms that have a considerable number of elements that remain in the same order as in the source