Restore index() functionality
The get_int_from_digits() method has been implemented in the CustomBaseNumber classes. This method is important in enabling the index of a term output by a PBTreeCombinatorialUnit to be looked up

Use int_npr() directly to set len_iter in Permutation class to increase limit on number of terms

As of this commit, all basic functionality has been restored. However, docstrings are currently no longer accurate and will be corrected in upcoming commits